FEM: constraint body heat flux, rename it to body heat source
