+++ src/Mod/Fem/App/CMakeLists.txt
@@ -106,7 +106,7 @@ SET(FemScripts_SRCS
 SET(FemObjectsScripts_SRCS
     PyObjects/__init__.py
     PyObjects/_FemConstraintSelfWeight.py
-    PyObjects/_FemConstraintBodyHeatFlux.py
+    PyObjects/_FemConstraintBodyHeatSource.py
     PyObjects/_FemConstraintFlowVelocity.py
     PyObjects/_FemConstraintInitialFlowVelocity.py
     PyObjects/_FemElementFluid1D.py
@@ -169,7 +169,7 @@ SET(FemGuiScripts_SRCS
     PyGui/__init__.py
     PyGui/_CommandFemAnalysis.py
     PyGui/_CommandFemConstraintSelfWeight.py
-    PyGui/_CommandFemConstraintBodyHeatFlux.py
+    PyGui/_CommandFemConstraintBodyHeatSource.py
     PyGui/_CommandFemConstraintFlowVelocity.py
     PyGui/_CommandFemConstraintInitialFlowVelocity.py
     PyGui/_CommandFemElementFluid1D.py
@@ -206,7 +206,7 @@ SET(FemGuiScripts_SRCS
     PyGui/_TaskPanelFemSolverCalculix.py
     PyGui/_TaskPanelFemSolverControl.py
     PyGui/_ViewProviderFemConstraintSelfWeight.py
-    PyGui/_ViewProviderFemConstraintBodyHeatFlux.py
+    PyGui/_ViewProviderFemConstraintBodyHeatSource.py
     PyGui/_ViewProviderFemConstraintFlowVelocity.py
     PyGui/_ViewProviderFemConstraintInitialFlowVelocity.py
     PyGui/_ViewProviderFemElementFluid1D.py
