+++ src/Mod/Fem/FemSolver/elmer/writer.py
@@ -251,7 +251,7 @@ class Writer(object):
             self._handled(obj)
 
     def _handleHeatBodyForces(self, bodies):
-        obj = self._getSingleMember("Fem::FemConstraintBodyHeatFlux")
+        obj = self._getSingleMember("Fem::FemConstraintBodyHeatSource")
         if obj is not None:
             for name in bodies:
                 heatSource = getFromUi(obj.HeatFlux, "W/kg", "L^2*T^-3")
